Prevent circular import by refactoring
